refactor: move walkValueForEntity out of shared
